Deduplicate the _withLocalState example tests
Each affected test now runs as two subtests: 'local', which always runs
against a temporary DIY backend, and 'service', which is skipped unless
PULUMI_TEST_USE_SERVICE=true. Previously the _withLocalState variants
were exact duplicates of their base tests on every PR and merge-queue
run, because ProgramTest already falls back to a temporary DIY backend
when PULUMI_TEST_USE_SERVICE is not "true".

examples_acceptance_test.go held only three of the duplicate tests plus
shared helpers; the helpers moved to examples_test.go and the file is
deleted. The filename had no special meaning: acceptance and
integration CI jobs run the same Go packages.
